More resilient kernel version checks
This allows the kernel version check code to use only a portion of the
string returned by `uname`, chopping off bits on the end until something
parseable is left.  As an example, the string `4.14.77-81.59.amzn2.x86_64`
is unparsable due to the underscore, and so with this change it will be
parsed as the version string `v"4.14.77-81.59.amzn2.x86"`.  This is
fine, as all we care about is the version number at the beginning.