Allow sparse GCC version support per architecture
Right now we assume that every GCC version is available for every
architecture. E.g. we always try to use the GCC 4.8 for auditing.
For Darwin Aarch64, we won't have any GCC shard smaller than 11.
We do already support falling back to the closest available version
if the preferred version is not supported. This simply extends
that check to only include compiler shards that we actually have
artifacts for.