Fix `clock_gettime` errors on old macOS machines (#134)
Our GCC shards did not correctly configure themselves to target the
oldest supported macOS build.  Instead, because GCC doesn't fully
understand how to use the availability macros that macOS uses, it
assumed that it was able to use `clock_gettime()`, which caused a
reliance on that symbol within `libgfortran` itself.  This resulted in
our CSL being unusable on macOS 10.11 and earlier.  This PR updates the
GCC shards [0] and also adds compiler flags to enforce that users who
attempt to use `gcc` on macOS do not end up making the same mistake.

[0] https://github.com/JuliaPackaging/Yggdrasil/pull/2836