[Runner] Improve compile-only flags checks and fix `objc` wrapper
The `objc` wrapper is currently roughly equivalent to `clang -x objective-c`,
the problem is that `-x <LANG>` can't be used when linking, otherwise the
compiler driver will assume that the input files are plain text source code
files.

As a simple rule, we assume that we are compiling (as opposed to linking) when
no argument is an object file (ending with `.o` extension) and make the `objc`
wrapper pass the `-x objective-c` flag only when compiling (and not linking),
instead of doing so unconditionally.