Reorganise Objective-C wrapper and Meson toolchain file
We can't make the Objective-C compiler always be the same as `clang -x
objective-c` because this won't work when using the compiler as linker: when
specifying the language with `-x <LANG>` the input files are assumed to always
be plain text source code files, but you pass object files to the linker.

This changeset removes the `objc` wrapper, which was created solely with the
goal to build Glib with Meson (which we never managed to do so far anyways) and
instead sets in the Meson toolchain file the Objective-C compiler to `cc` and
passes `-x objective-c` only when compiling but not linking.  This is more
correct and less error-prone than what we have now.

We also automatically add `-I${includedir}` and `-L{libdir}` for all languages
in the Meson toolchain file, to avoid having to manually patch this file during
builds, as Meson doesn't read environment variables when doing
cross-compilation.