[Runner] Generate CSL paths only for target and host platforms
Having all Intel Linux platforms in the CSL paths was a mess hard to deal with.
The simplest thing to do is to generate the paths only for target and host, and
sort them correctly: if they have same architecture, have the host first, if the
have different architectures, have the target first.