[Runner] Move `csl_paths(host_platform)` at the beginning of `LD_LIBRARY_PATH` (#157)
* [Runner] Fix CSL paths

* [Runner] Move `csl_paths(host_platform)` at the beginning of `LD_LIBRARY_PATH`

This should make it easier to run applications we build with our toolchain.

* [Rootfs] Update to fix CSL libraries

* [Runner] Override `LD_LIBRARY_PATH` inside the compiler wrappers

* [Runner] Generate CSL paths only for target and host platforms

Having all Intel Linux platforms in the CSL paths was a mess hard to deal with.
The simplest thing to do is to generate the paths only for target and host, and
sort them correctly: if they have same architecture, have the host first, if the
have different architectures, have the target first.

* Simplify reading of some buffers in tests

* [Runner] Add helper function to set `LD_LIBRARY_PATH`

* [Runner] Have always `host` first in CSL paths for `LD_LIBRARY_PATH`

Having `target` first would make C++ programs for `i686-linux-musl` work, but
then we can't run C++ programs for the host.  We really need to fix the musl
loader.