[BuildToolchains] Use `<aatriplet>-cc` as linker in Cargo config
I don't understand why we were forcing `gcc`, on `aarch64-apple-darwin` this
causes lots of trouble, because Cargo would pass in the argument `-arch arm64`
which isn't accepted by GCC, but it is by LLVM.

Now that we can JIT-generate the file, write the config only for host and target
platforms, we don't have the toolchains for the other platforms anyway.

Additionally, remove some workarounds previously added in `Runner.jl` to fix
errors in the old static config file.