Add `CrossPlatform` type
This type allows us to serialize and reason about a platform that
contains `host` and `target` sub-platforms; to represent our cross
compiler artifacts such as `GCC_jll` or even `LinuxKernelHeaders_jll`.

In the case of `GCC_jll` we will generally have fully-specified `host`
and `target` sub-platforms: e.g. that binary artifact will run on
exactly `x86_64-linux-musl`, and it will target exactly
`aarch64-apple-darwin14`.

In the case of `LinuxKernelHeaders_jll`, because it consists of simple
header files, it's actually a "host-any" artifact, and should be
represented by having `host` be an `AnyPlatform`, as it can be used on
any host, but still is target-specific.

In the case of `Clang_jll`, we may build an LLVM that can target any of
our valid targets, in which case we build a "target-any" artifact,
however it can only be executed on the host platform it was built for.

This `CrossPlatform` type can represent all of these situations, and
will allow us to use the typical `platforms_match()`,
`select_downloadable_artifacts()`, etc... `Artifacts.jl` machinery that
we know and love.  The ideal future state is one in which we can simply
do something like:

```
using BinaryBuilderBase: get_gcc_artifact

gcc_artifact = get_gcc_artifact(
    Platform("x86_64", "linux") => Platform("aarch64", "macos")
)
```

And it will just use `GCC_jll`, `CrossPlatform` and
`select_downloadable_artifact()` internally as would be expected.