Fix some issues reported on Slack (#243)
* Fix `FileSource` incorrectly trying to replace `srcdir`

Sometimes, a user tries to download a file that is an archive but it
accidentally gets seen as a `FileSource` due to our imperfect archive
detection. Rather than erroring out here with a bad `cp()` invocation,
automatically plop the file into `srcdir`, and allow the build to
continue.

* Check `path` instead of `url` for `archive_extensions`

This allows us to perform a bit of URL cleaning to drop things like
`?raw=true` at the end of URLs from GitHub, for instance.  Best to check
the given destination path name, rather than the original URL name.

* Update sources.jl