Remove GCCBootstrap artifacts for FreeBSD 12.2 (#315)
They seem to be picked up by Yggdrasil CI rather than the 13.2
artifacts, but the 13.2 artifacts are the ones that should be used.

Note that the 12.2 artifacts for PlatformSupport are left untouched
because in that case the 12.2 and 13.2 artifacts correspond to distinct
versions.

The minor version is bumped here, even though this is technically
breaking for users on FreeBSD 12. However, FreeBSD 12 is past EOL so I
think it's reasonable to make it a minor rather than major change.