[RootFS] Use GCC 13+ for riscv64
We've had a few problems with GCC 12 on riscv64 (missing `riscv_vector.h` header
file, broken support for atomics) that we're often forced to use GCC 13.  Let's
always use GCC 13 at minimum and live happier.

Once we get GCC 14 in we may bump the minimum riscv64 GCC toolchain to that
version, because it introduces better support for the version 1.0 of the vector
extension, important for OpenBLAS and other software.