Update FreeBSD versions to 13.4 on x86-64 and 14.1 on AArch64 (#418)
* Update PlatformSupport to v2025.02.15

Notably updates FreeBSD to 13.4 on x86-64 and 14.1 on AArch64.

* Artifacts for GCC 9-14 on 14.1 AArch64, 4-14 on 13.4 x86-64

We were able to build GCC 7 and 8 on FreeBSD 13.2 AArch64, but neither
builds successfully with 14.1. For both 7 and 8, an error about the file
format of `crtn.o` is thrown from `collect2`/`ld`. It's unclear why
that's happening and I'm not sure I care enough to figure it out and fix
it vs. just bumping the minimum GCC version on FreeBSD AArch64 to 9.
(That is what I did.)

* Bump package version for new release